requirements.txt
tweepy==4.14.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==1.10.13
openai>=1.0.0
tenacity==8.2.3
streamlit==1.31.0
pytest==7.4.3
pytest-asyncio==0.23.2
sqlalchemy==2.0.25
aiosqlite==0.19.0
jinja2==3.1.2
